🚚 Move: `SectionTitle.svelte` -> `$lib/components/`
