+++ src/lib/components/SectionTitle.svelte
